Port to v.2022.4.62.1 of tMod
Ft: Very scuffed entity sources, someone'll need to get around to assigning proper sources

+++ Content/Items/Weapons/Ranged/HauntedEarth.cs
@@ -38,7 +38,7 @@ namespace DestinyMod.Content.Items.Weapons.Ranged
 			return false;
 		}
 
-		public override void ModifyWeaponCrit(Player player, ref int crit)
+		public override void ModifyWeaponCrit(Player player, ref float crit)
 		{
 			if (player.velocity == Vector2.Zero)
 			{
